feat(coinmarket): refactor - make independent types
