+++ packages/suite/src/types/coinmarket/coinmarket.ts
@@ -1,0 +1,83 @@
+import type { BuyInfo } from 'src/actions/wallet/coinmarketBuyActions';
+import type { ExchangeInfo } from 'src/actions/wallet/coinmarketExchangeActions';
+import type { SellInfo } from 'src/actions/wallet/coinmarketSellActions';
+import type {
+    Trade,
+    TradeBuy,
+    TradeExchange,
+    TradeSell,
+    TradeType,
+} from 'src/types/wallet/coinmarketCommonTypes';
+import {
+    BuyTradeStatus,
+    ExchangeTradeStatus,
+    SavingsTradeItemStatus,
+    SellTradeStatus,
+    WatchBuyTradeResponse,
+    WatchExchangeTradeResponse,
+    WatchSellTradeResponse,
+} from 'invity-api';
+import { Account } from '@suite-common/wallet-types';
+import { AnyAction, Dispatch } from 'redux';
+import { State } from 'src/reducers/wallet/coinmarketReducer';
+
+export type CoinmarketTradeBuyType = 'buy';
+export type CoinmarketTradeSellType = 'sell';
+export type CoinmarketTradeExchangeType = 'exchange';
+export type CoinmarketTradeType =
+    | CoinmarketTradeBuyType
+    | CoinmarketTradeSellType
+    | CoinmarketTradeExchangeType;
+
+export type CoinmarketTradeMapProps = {
+    buy: TradeBuy;
+    sell: TradeSell;
+    exchange: TradeExchange;
+};
+
+export type CoinmarketTradeInfoMapProps = {
+    buy: BuyInfo;
+    sell: SellInfo;
+    exchange: ExchangeInfo;
+};
+
+export type CoinmarketWatchTradeResponseMapProps = {
+    buy: WatchBuyTradeResponse;
+    sell: WatchSellTradeResponse;
+    exchange: WatchExchangeTradeResponse;
+};
+
+export interface CoinmarketGetTypedTradeProps {
+    trades: Trade[];
+    tradeType: CoinmarketTradeType;
+    transactionId: string | undefined;
+}
+
+export type CoinmarketTradeStatusType =
+    | BuyTradeStatus
+    | SellTradeStatus
+    | ExchangeTradeStatus
+    | SavingsTradeItemStatus;
+
+export interface CoinmarketGetDetailDataProps {
+    coinmarket: State;
+    tradeType: TradeType;
+}
+
+export interface CoinmarketGetTypedInfoTradeProps {
+    coinmarket: State;
+    tradeType: CoinmarketTradeType;
+}
+
+export interface CoinmarketUseWatchTradeProps<T extends CoinmarketTradeType> {
+    account: Account | undefined;
+    trade: CoinmarketTradeMapProps[T] | undefined;
+}
+
+export interface CoinmarketWatchTradeProps<T extends CoinmarketTradeType> {
+    trade: CoinmarketTradeMapProps[T];
+    account: Account;
+    refreshCount: number;
+    dispatch: Dispatch<AnyAction>;
+    removeDraft: (key: string) => void;
+}
